Fix propose accept scenario
Daml model is mostly unchanged except for the addition of the alias
template.

I upgraded the UI code by taking the 2.0.0 version of create-daml-app
and adding back the custom typescript code from this scenario adding
the required changes to pick up aliases as needed. Tested locally that
this works as expected.